Tests prepared to be continued
